fix: pre-existing win_unit_tests build errors
- BootstrapResult: add explicit destructor for unique_ptr<DnnSsdFaceDetector> incomplete type
- FaceDatabase: cast int64_t count to int for OpenCV FileStorage serialization

+++ src/win/src/FaceDatabase.cpp
@@ -66,7 +66,7 @@ bool FaceDatabase::save(const std::filesystem::path& path) const {
         std::copy(e.mean.begin(), e.mean.end(), meanMat.ptr<float>());
         fs << "{";
         fs << "id" << e.id;
-        fs << "count" << e.count;
+        fs << "count" << static_cast<int>(e.count);
         fs << "mean" << meanMat;
         fs << "}";
     }
